stm32: add kernel patch to switch wifi SDIO in In Band interrupt mode
An issue has been discovered on the wifi module during iperf3 tests, after
few seconds the wifi module gets stuck. And the only way to recover it is
to reset the board.

The issue seems to be related to the SDIO Out Of Band interrupt
implementation for the wifi. A workaround is to pass from Out Of Band
interrupt to In Band interrupt.

This kernel patch is only a part of the workaround. The NVRAM configuration
file of the wifi module shall be modified too.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>