.github/workflows: clean packageinfo of linux kernel before defconfig
This patch is a workaround to force the parsing of linux kernel package,
after the installation of the stm32 target.
The linux kernel package is scanned before the installation of the stm32
target. So the stm32 kernel module makefile
(target/linux/feeds/stm32/modules.mk) is not parsed because it doesn't
exists, consequently the package kmod-usb-stm32-usbh doesn't exists.

Awaiting an upstream patch, linux kernel packageinfo is deleted before to
call defconfig target to force the parsing of the linux kernel package.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>